refactor: 폴더구조 정리
- component
	- @shared : 범용적으로 많이 쓰일만한 컴포넌트
	- Cart: cart에 대한 컴포넌트
	- Product: Product에 대한 컴포넌트
