+++ src/components/@shared/ShoppingCartIcon/index.tsx
