+++ src/components/@shared/Skeleton/index.styles.ts
