+++ src/components/Cart/CartItem/index.styles.ts
