+++ src/components/Product/ItemDetails/index.styles.ts
