+++ src/components/Product/ItemSkeleton/index.stories.jsx
