+++ src/components/Product/ItemSkeleton/index.styles.ts
