+++ src/modules/type.ts
@@ -1,6 +1,6 @@
 import { AnyAction } from "redux";
-import { RootState } from "..";
-import { ThunkAction } from "../../lib/thunk.type";
+import { RootState } from ".";
+import { ThunkAction } from "../lib/thunk.type";
 
 export type AppThunk<ReturnType = void> = ThunkAction<
   ReturnType,
